⚠ Correctly handle read_count
The read syscall can return smaller non-zero values in a number of cases,
such as when the process is interrupted or we are reading from a pipe.
An exact zero, however, means that this is (at least currently)
the end of the file.

This was trickier, since BUFFER_LEN is not just some arbitrary io buffer,
but instead determines the signed blocks ("messages") of the authenticated encryption.

One consequence of the current implementation is that if the plaintext
is exactly a multiple of BUFFER_LEN, we will produce an empty message
at the end.
This is not such a big overhead, just an extra tag of 16 bytes, but
could be avoided by lookahead with two buffers or a larger buffer.

I also saw the correct +16 bytes in the article, but they were missing here.

Also added a basic roundtrip integration test, which now passes.